+++ BitTorrent/BitTorrentLibrary/Tracker.cs
@@ -45,17 +45,17 @@ namespace BitTorrentLibrary
         internal AnnounceResponse lastResponse;                 // Last announce response 
         internal TrackerEvent Event { get; set; }               // Current state of torrent downloading
         public string PeerID { get; }                           // Peers unique ID
-        public uint Port { get; } = Host.DefaultPort;           // Port that client s listening on 
+        public int Port { get; } = Host.DefaultPort;           // Port that client s listening on 
         public string Ip { get; set; }                          // IP of host performing announce
-        public uint Compact { get; } = 1;                       // Is the returned peer list compressed (1=yes,0=no)
-        public uint NoPeerID { get; }                           // Unique peer ID for downloader
+        public int Compact { get; } = 1;                       // Is the returned peer list compressed (1=yes,0=no)
+        public int NoPeerID { get; }                           // Unique peer ID for downloader
         public string Key { get; }                              // An additional identification that is not shared with any other peers (optional)
         public string TrackerID { get; set; }                   // String that the client should send back on its next announcements. (optional).
         public int NumWanted { get; set; } = 5;                 // Number of required download clients
         public byte[] InfoHash { get; }                         // Encoded info hash for URI
         public string TrackerURL { get; }                       // Tracker URL
-        public uint Interval { get; set; } = 2000;              // Polling interval between each announce
-        public uint MinInterval { get; set; }                   // Minumum allowed polling interval
+        public int Interval { get; set; } = 2000;              // Polling interval between each announce
+        public int MinInterval { get; set; }                   // Minumum allowed polling interval
         public TrackerCallBack CallBack { get; set; }           // Tracker ping callback function
         public Object CallBackData { get; set; }                // Tracker ping callback function data
         public UInt64 Downloaded => _tc.TotalBytesDownloaded;   // Total downloaded bytes of torrent to local client
@@ -89,7 +89,7 @@ namespace BitTorrentLibrary
                         Log.Logger.Info("(Tracker) Queuing new peers for swarm ....");
                         if ((tracker._tc.Status == TorrentStatus.Downloading)&&(tracker.peerSwarmQueue.Count==0))
                         {
-                            UInt32 peerThreshHold = (UInt32)tracker._tc.maximumSwarmSize;
+                            int peerThreshHold = tracker._tc.maximumSwarmSize;
                             foreach (var peerDetails in tracker.lastResponse.peers)
                             {
                                 if (!tracker._tc.manager.IsPeerDead(peerDetails.ip))
@@ -133,7 +133,7 @@ namespace BitTorrentLibrary
             {
                 if (response.interval > MinInterval)
                 {
-                    UInt32 oldInterval = Interval;
+                    int oldInterval = Interval;
                     Interval = response.interval;
                     if (oldInterval != Interval)
                     {
@@ -276,7 +276,7 @@ namespace BitTorrentLibrary
         /// Set tracker announce interval for when seeding.
         /// </summary>
         /// <param name="seedingInerval"></param>
-        public void SetSeedingInterval(UInt32 seedingInerval)
+        public void SetSeedingInterval(int seedingInerval)
         {
             if (_tc.Status == TorrentStatus.Seeding)
             {
